requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
distro==1.9.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==1.24.3
openai==1.12.0
openai-whisper==20240930
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.0
regex==2024.11.6
requests==2.32.3
scipy==1.10.1
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
